[Feat, Style] 모델 파일 분리 및 Order, RockPaperScissors 클래스 생성, makeComputerHandSign 함수 기능 구현
