+++ RockPaperScissors/RockPaperScissors/Models/Order.swift
@@ -1,0 +1,15 @@
+//
+//  Order.swift
+//  RockPaperScissors
+//
+//  Created by Swain Yun on 2023/09/06.
+//
+
+import Foundation
+
+enum Order: Int, CaseIterable {
+    case exitProgram = 0
+    case scissors = 1
+    case rock = 2
+    case paper = 3
+}
